backend/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==5.3.0
altgraph==0.17.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.8.0
arabic-reshaper==3.0.0
asttokens==2.4.1
astunparse==1.6.3
attrs==24.2.0
autots==1.0.0
beautifulsoup4==4.12.3
blinker==1.8.2
blis==1.2.0
branca==0.8.1
cachetools==5.4.0
camel_tools==1.5.5
catalogue==2.0.10
certifi==2024.7.4
cffi==2.0.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.20.0
cmake==3.30.2
colorama==0.4.6
comm==0.2.2
comtypes==1.4.7
confection==0.1.5
contourpy==1.2.1
cryptography==47.0.0
cycler==0.12.1
cymem==2.0.11
Cython==3.0.11
debugpy==1.8.5
decorator==5.1.1
deepface-cv2==0.0.1
dill==0.3.9
distlib==0.3.8
distro==1.9.0
dlib==19.24.5
dnspython==2.7.0
docopt==0.6.2
editdistance==0.8.1
emoji==2.14.1
executing==2.1.0
face-recognition==1.3.0
face_recognition_models==0.3.0
Faker==37.4.2
Farasa==0.0.1
fastapi==0.136.1
fastjsonschema==2.21.2
filelock==3.15.4
fire==0.6.0
Flask==3.0.3
flatbuffers==24.3.25
folium==0.19.4
fonttools==4.53.1
frozenlist==1.5.0
fsspec==2025.2.0
future==1.0.0
gast==0.6.0
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.49.2
google-auth-httplib2==0.2.0
google-genai==1.74.0
google-generativeai==0.8.6
google-pasta==0.2.0
googleapis-common-protos==1.66.0
greenlet==3.5.0
grpcio==1.70.0
grpcio-status==1.70.0
gTTS==2.5.2
gunicorn==22.0.0
h11==0.16.0
h5py==3.11.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.28.1
idna==3.7
imbalanced-learn==0.13.0
ipykernel==6.29.5
ipython==8.27.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.14.0
joblib==1.4.2
Js2Py==0.74
jsonpath-python==1.1.5
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.4.1
kiwisolver==1.4.5
labelImg==1.8.6
langcodes==3.5.0
language_data==1.3.0
libclang==18.1.1
lxml==5.3.0
marisa-trie==1.2.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.0
MouseInfo==0.1.3
mpmath==1.3.0
muddler==0.1.3
multidict==6.1.0
murmurhash==1.0.12
mysql-connector-python==9.4.0
namex==0.0.8
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
openai==2.33.0
opencv-python==4.10.0.84
openrouter==0.9.1
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.2.2
parso==0.8.4
patsy==1.0.2
pefile==2023.2.7
pillow==10.4.0
pipwin==0.5.2
platformdirs==4.2.2
playsound==1.3.0
polars==1.39.3
polars-runtime-32==1.39.3
preshed==3.0.9
prompt_toolkit==3.0.47
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.6
psutil==6.0.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
pydeck==0.9.1
pydub==0.25.1
pygame==2.6.0
PyGetWindow==0.0.9
Pygments==2.18.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2025.1
pyjsparser==2.7.1
pymongo==4.11
PyMsgBox==1.0.9
pyparsing==3.1.2
pyperclip==1.9.0
pypiwin32==223
PyPrind==2.11.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
PyQt6==6.8.0
PyQt6-Qt6==6.8.1
PyQt6-WebEngine==6.8.0
PyQt6-WebEngine-Qt6==6.8.1
PyQt6_sip==13.9.1
PyRect==0.2.0
pyrsistent==0.20.0
PyScreeze==1.0.1
pySmartDL==1.3.4
PySocks==1.7.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-telegram-bot==22.4
pyttsx3==2.97
pytweening==1.2.0
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.7.1
rpds-py==0.20.0
rsa==4.9
safetensors==0.5.2
scikit-fuzzy==0.5.0
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
setuptools==72.1.0
shellingham==1.5.4
six==1.16.0
sklearn-compat==0.1.3
smart-open==7.1.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.10.4
split-folders==0.5.1
SQLAlchemy==2.0.49
srsly==2.5.1
stack-data==0.6.3
starlette==1.0.0
statsmodels==0.14.5
streamlit==1.37.1
sympy==1.13.1
tabulate==0.9.0
tenacity==8.5.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
termcolor==2.4.0
thinc==8.3.4
threadpoolctl==3.5.0
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.43.4
typer==0.15.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.46.0
virtualenv==20.26.3
wasabi==1.1.3
watchdog==4.0.1
wcwidth==0.2.13
weasel==0.4.1
websockets==16.0
Werkzeug==3.0.3
wheel==0.44.0
Wikipedia-API==0.8.1
wrapt==1.16.0
xgboost @ file:///C:/Users/victus/Downloads/xgboost-3.0.2-py3-none-win_amd64.whl#sha256=b4c89b71d134da9fa6318e3c9f5459317d1013b4d57059d10ed2840750e2f7e1
xyzservices==2025.1.0
yarl==1.18.3
